@@ -1,20 +1,19 @@
 import json
 import os.path
-import yaml
 
+import yaml
 from chaoslib.exceptions import ActivityFailed
 from chaoslib.types import Secrets
 from kubernetes import client
-from logzero import logger
 from kubernetes.client.rest import ApiException
+from logzero import logger
 
 from chaosk8s import create_k8s_api_client
 
 __all__ = ["create_deployment", "delete_deployment", "scale_deployment"]
 
 
-def create_deployment(spec_path: str, ns: str = "default",
-                      secrets: Secrets = None):
+def create_deployment(spec_path: str, ns: str = "default", secrets: Secrets = None):
     """
     Create a deployment described by the deployment config, which must be the
     path to the JSON or YAML representation of the deployment.
@@ -23,20 +22,23 @@ def create_deployment(spec_path: str, ns: str = "default",
 
     with open(spec_path) as f:
         p, ext = os.path.splitext(spec_path)
-        if ext == '.json':
+        if ext == ".json":
             deployment = json.loads(f.read())
-        elif ext in ['.yml', '.yaml']:
+        elif ext in [".yml", ".yaml"]:
             deployment = yaml.load(f.read())
         else:
-            raise ActivityFailed(
-                "cannot process {path}".format(path=spec_path))
+            raise ActivityFailed(f"cannot process {spec_path}")
 
     v1 = client.AppsV1Api(api)
-    resp = v1.create_namespaced_deployment(ns, body=deployment)
+    _ = v1.create_namespaced_deployment(ns, body=deployment)
 
 
-def delete_deployment(name: str = None, ns: str = "default",
-                      label_selector: str = None, secrets: Secrets = None):
+def delete_deployment(
+    name: str = None,
+    ns: str = "default",
+    label_selector: str = None,
+    secrets: Secrets = None,
+):
     """
     Delete a deployment by `name` or `label_selector` in the namespace `ns`.
 
@@ -51,23 +53,22 @@ def delete_deployment(name: str = None, ns: str = "default",
     v1 = client.AppsV1Api(api)
 
     if name:
-        ret = v1.list_namespaced_deployment(
-            ns, field_selector="metadata.name={}".format(name))
+        ret = v1.list_namespaced_deployment(ns, field_selector=f"metadata.name={name}")
     elif label_selector:
         ret = v1.list_namespaced_deployment(ns, label_selector=label_selector)
     else:
         ret = v1.list_namespaced_deployment(ns)
 
-    logger.debug("Found {d} deployments named '{n}'".format(
-        d=len(ret.items), n=name))
+    logger.debug(f"Found {len(ret.items)} deployments named '{name}'")
 
     body = client.V1DeleteOptions()
     for d in ret.items:
         v1.delete_namespaced_deployment(d.metadata.name, ns, body=body)
 
 
-def scale_deployment(name: str, replicas: int, ns: str = "default",
-                     secrets: Secrets = None):
+def scale_deployment(
+    name: str, replicas: int, ns: str = "default", secrets: Secrets = None
+):
     """
     Scale a deployment up or down. The `name` is the name of the deployment.
     """
@@ -79,5 +80,5 @@ def scale_deployment(name: str, replicas: int, ns: str = "default",
         v1.patch_namespaced_deployment(name=name, namespace=ns, body=body)
     except ApiException as e:
         raise ActivityFailed(
-            "failed to scale '{s}' to {r} replicas: {e}".format(
-                s=name, r=replicas, e=str(e)))
+            f"failed to scale '{name}' to {replicas} replicas: {str(e)}"
+        )
