@@ -1,6 +1,5 @@
-# -*- coding: utf-8 -*-
 from datetime import datetime
-from typing import Dict, Union, List
+from typing import Dict, List, Union
 
 import dateparser
 from chaoslib.exceptions import ActivityFailed
@@ -16,15 +15,19 @@ __all__ = [
     "pods_not_in_phase",
     "read_pod_logs",
     "count_pods",
-    "pod_is_not_available"
+    "pod_is_not_available",
 ]
 
 
-def read_pod_logs(name: str = None, last: Union[str, None] = None,
-                  ns: str = "default", from_previous: bool = False,
-                  label_selector: str = "name in ({name})",
-                  container_name: str = None,
-                  secrets: Secrets = None) -> Dict[str, str]:
+def read_pod_logs(
+    name: str = None,
+    last: Union[str, None] = None,
+    ns: str = "default",
+    from_previous: bool = False,
+    label_selector: str = "name in ({name})",
+    container_name: str = None,
+    secrets: Secrets = None,
+) -> Dict[str, str]:
     """
     Fetch logs for all the pods with the label `"name"` set to `name` and
     return a dictionary with the keys being the pod's name and the values
@@ -50,9 +53,10 @@ def read_pod_logs(name: str = None, last: Union[str, None] = None,
     else:
         ret = v1.list_namespaced_pod(ns)
 
-    logger.debug("Found {d} pods: [{p}] in ns '{n}'".format(
-        d=len(ret.items), n=ns,
-        p=', '.join([p.metadata.name for p in ret.items])))
+    logger.debug(
+        f"Found {len(ret.items)} "
+        f"pods: [{', '.join([p.metadata.name for p in ret.items])}] in ns '{ns}'"
+    )
 
     since = None
     if last:
@@ -65,7 +69,7 @@ def read_pod_logs(name: str = None, last: Union[str, None] = None,
         previous=from_previous,
         timestamps=True,
         container=container_name or "",  # None is not a valid value
-        _preload_content=False
+        _preload_content=False,
     )
 
     if since:
@@ -74,15 +78,19 @@ def read_pod_logs(name: str = None, last: Union[str, None] = None,
     logs = {}
     for p in ret.items:
         name = p.metadata.name
-        logger.debug("Fetching logs for pod '{n}'".format(n=name))
+        logger.debug(f"Fetching logs for pod '{name}'")
         r = v1.read_namespaced_pod_log(name, **params)
-        logs[name] = r.read().decode('utf-8')
+        logs[name] = r.read().decode("utf-8")
 
     return logs
 
 
-def pods_in_phase(label_selector: str, phase: str = "Running",
-                  ns: str = "default", secrets: Secrets = None) -> bool:
+def pods_in_phase(
+    label_selector: str,
+    phase: str = "Running",
+    ns: str = "default",
+    secrets: Secrets = None,
+) -> bool:
     """
     Lookup a pod by `label_selector` in the namespace `ns`.
 
@@ -94,28 +102,33 @@ def pods_in_phase(label_selector: str, phase: str = "Running",
     v1 = client.CoreV1Api(api)
     if label_selector:
         ret = v1.list_namespaced_pod(ns, label_selector=label_selector)
-        logger.debug("Found {d} pods matching label '{n}' in ns '{s}'".format(
-            d=len(ret.items), n=label_selector, s=ns))
+        logger.debug(
+            f"Found {len(ret.items)} pods matching label '{label_selector}'"
+            f" in ns '{ns}'"
+        )
     else:
         ret = v1.list_namespaced_pod(ns)
-        logger.debug("Found {d} pods in ns '{n}'".format(
-            d=len(ret.items), n=ns))
+        logger.debug(f"Found {len(ret.items)} pods in ns '{ns}'")
 
     if not ret.items:
-        raise ActivityFailed(
-            "no pods '{name}' were found".format(name=label_selector))
+        raise ActivityFailed(f"no pods '{label_selector}' were found")
 
     for d in ret.items:
         if d.status.phase != phase:
             raise ActivityFailed(
-                "pod '{name}' is in phase '{s}' but should be '{p}'".format(
-                    name=label_selector, s=d.status.phase, p=phase))
+                f"pod '{label_selector}' is in phase '{d.status.phase}'"
+                f" but should be '{phase}'"
+            )
 
     return True
 
 
-def pods_in_conditions(label_selector: str, conditions: List[Dict[str, str]],
-                       ns: str = "default", secrets: Secrets = None) -> bool:
+def pods_in_conditions(
+    label_selector: str,
+    conditions: List[Dict[str, str]],
+    ns: str = "default",
+    secrets: Secrets = None,
+) -> bool:
     """
     Lookup a pod by `label_selector` in the namespace `ns`.
 
@@ -127,36 +140,38 @@ def pods_in_conditions(label_selector: str, conditions: List[Dict[str, str]],
     v1 = client.CoreV1Api(api)
     if label_selector:
         ret = v1.list_namespaced_pod(ns, label_selector=label_selector)
-        logger.debug("Found {d} pods matching label '{n}' in ns '{s}'".format(
-            d=len(ret.items), n=label_selector, s=ns))
+        logger.debug(
+            f"Found {len(ret.items)} pods matching label '{label_selector}'"
+            f" in ns '{ns}'"
+        )
     else:
         ret = v1.list_namespaced_pod(ns)
-        logger.debug("Found {d} pods in ns '{n}'".format(
-            d=len(ret.items), n=ns))
+        logger.debug(f"Found {len(ret.items)} pods in ns '{ns}'")
 
     if not ret.items:
-        raise ActivityFailed(
-            "no pods '{name}' were found".format(name=label_selector))
+        raise ActivityFailed(f"no pods '{label_selector}' were found")
 
     for d in ret.items:
         # create a list of hash to compare with the given conditions
         pod_conditions = [
-            {"type": pc.type, "status": pc.status}
-            for pc in d.status.conditions
+            {"type": pc.type, "status": pc.status} for pc in d.status.conditions
         ]
         for condition in conditions:
             if condition not in pod_conditions:
-                raise ActivityFailed("pod {name} does not match the following "
-                                     "given condition: {condition}".format(
-                                         name=d.metadata.name,
-                                         condition=condition
-                                     ))
+                raise ActivityFailed(
+                    f"pod {d.metadata.name} does not match the following "
+                    f"given condition: {condition}"
+                )
 
     return True
 
 
-def pods_not_in_phase(label_selector: str, phase: str = "Running",
-                      ns: str = "default", secrets: Secrets = None) -> bool:
+def pods_not_in_phase(
+    label_selector: str,
+    phase: str = "Running",
+    ns: str = "default",
+    secrets: Secrets = None,
+) -> bool:
     """
     Lookup a pod by `label_selector` in the namespace `ns`.
 
@@ -168,28 +183,29 @@ def pods_not_in_phase(label_selector: str, phase: str = "Running",
     v1 = client.CoreV1Api(api)
     if label_selector:
         ret = v1.list_namespaced_pod(ns, label_selector=label_selector)
-        logger.debug("Found {d} pods matching label '{n}' in ns '{s}'".format(
-            d=len(ret.items), n=label_selector, s=ns))
+        logger.debug(
+            f"Found {len(ret.items)} pods matching label '{label_selector}'"
+            f" in ns '{ns}'"
+        )
     else:
         ret = v1.list_namespaced_pod(ns)
-        logger.debug("Found {d} pods in ns '{n}'".format(
-            d=len(ret.items), n=ns))
+        logger.debug(f"Found {len(ret.items)} pods in ns '{ns}'")
 
     if not ret.items:
-        raise ActivityFailed(
-            "no pods '{name}' were found".format(name=label_selector))
+        raise ActivityFailed(f"no pods '{label_selector}' were found")
 
     for d in ret.items:
         if d.status.phase == phase:
             raise ActivityFailed(
-                "pod '{name}' should not be in phase '{s}'".format(
-                    name=label_selector, s=d.status.phase))
+                f"pod '{label_selector}' should not be in phase '{d.status.phase}'"
+            )
 
     return True
 
 
-def count_pods(label_selector: str, phase: str = None,
-               ns: str = "default", secrets: Secrets = None) -> int:
+def count_pods(
+    label_selector: str, phase: str = None, ns: str = "default", secrets: Secrets = None
+) -> int:
     """
     Count the number of pods matching the given selector in a given `phase`, if
     one is given.
@@ -199,12 +215,13 @@ def count_pods(label_selector: str, phase: str = None,
     v1 = client.CoreV1Api(api)
     if label_selector:
         ret = v1.list_namespaced_pod(ns, label_selector=label_selector)
-        logger.debug("Found {d} pods matching label '{n}' in ns '{s}'".format(
-            d=len(ret.items), n=label_selector, s=ns))
+        logger.debug(
+            f"Found {len(ret.items)} pods matching label '{label_selector}'"
+            f" in ns '{ns}'"
+        )
     else:
         ret = v1.list_namespaced_pod(ns)
-        logger.debug("Found {d} pods in ns '{n}'".format(
-            d=len(ret.items), n=ns))
+        logger.debug(f"Found {len(ret.items)} pods in ns '{ns}'")
 
     if not ret.items:
         return 0
@@ -220,9 +237,12 @@ def count_pods(label_selector: str, phase: str = None,
     return count
 
 
-def pod_is_not_available(name: str, ns: str = "default",
-                         label_selector: str = "name in ({name})",
-                         secrets: Secrets = None) -> bool:
+def pod_is_not_available(
+    name: str,
+    ns: str = "default",
+    label_selector: str = "name in ({name})",
+    secrets: Secrets = None,
+) -> bool:
     """
     Lookup pods with a `name` label set to the given `name` in the specified
     `ns`.
@@ -239,22 +259,20 @@ def pod_is_not_available(name: str, ns: str = "default",
     else:
         ret = v1.list_namespaced_pod(ns)
 
-    logger.debug("Found {d} pod(s) named '{n}' in ns '{s}".format(
-        d=len(ret.items), n=name, s=ns))
+    logger.debug(f"Found {len(ret.items)} pod(s) named '{name}' in ns '{ns}")
 
     for p in ret.items:
         phase = p.status.phase
-        logger.debug("Pod '{p}' has status '{s}'".format(
-            p=p.metadata.name, s=phase))
+        logger.debug(f"Pod '{p.metadata.name}' has status '{phase}'")
         if phase == "Running":
-            raise ActivityFailed(
-                "pod '{name}' is actually running".format(name=name))
+            raise ActivityFailed(f"pod '{name}' is actually running")
 
     return True
 
 
-def all_pods_healthy(ns: str = "default",
-                     secrets: Secrets = None) -> MicroservicesStatus:
+def all_pods_healthy(
+    ns: str = "default", secrets: Secrets = None
+) -> MicroservicesStatus:
     """
     Check all pods in the system are running and available.
 
@@ -274,8 +292,7 @@ def all_pods_healthy(ns: str = "default",
         elif phase not in ("Running", "Succeeded"):
             not_ready.append(p)
 
-    logger.debug("Found {d} failed and {n} not ready pods".format(
-        d=len(failed), n=len(not_ready)))
+    logger.debug(f"Found {len(failed)} failed and {len(not_ready)} not ready pods")
 
     # we probably should list them in the message
     if failed or not_ready:
