@@ -1,16 +1,18 @@
-# -*- coding: utf-8 -*-
-from unittest.mock import MagicMock, patch, ANY
+from unittest.mock import ANY, MagicMock, patch
 
 import pytest
 from chaoslib.exceptions import ActivityFailed
 
-from chaosk8s.statefulset.actions import scale_statefulset, \
-    remove_statefulset, create_statefulset
+from chaosk8s.statefulset.actions import (
+    create_statefulset,
+    remove_statefulset,
+    scale_statefulset,
+)
 
 
-@patch('chaosk8s.has_local_config_file', autospec=True)
-@patch('chaosk8s.statefulset.actions.client', autospec=True)
-@patch('chaosk8s.client')
+@patch("chaosk8s.has_local_config_file", autospec=True)
+@patch("chaosk8s.statefulset.actions.client", autospec=True)
+@patch("chaosk8s.client")
 def test_scale_statefulset(cl, client, has_conf):
     has_conf.return_value = False
 
@@ -23,12 +25,13 @@ def test_scale_statefulset(cl, client, has_conf):
 
     assert v1.patch_namespaced_stateful_set.call_count == 1
     v1.patch_namespaced_stateful_set.assert_called_with(
-        "my-statefulset", namespace="default", body=body)
+        "my-statefulset", namespace="default", body=body
+    )
 
 
-@patch('chaosk8s.has_local_config_file', autospec=True)
-@patch('chaosk8s.statefulset.actions.client', autospec=True)
-@patch('chaosk8s.client')
+@patch("chaosk8s.has_local_config_file", autospec=True)
+@patch("chaosk8s.statefulset.actions.client", autospec=True)
+@patch("chaosk8s.client")
 def test_removing_statefulset_with_name(cl, client, has_conf):
     has_conf.return_value = False
 
@@ -43,15 +46,17 @@ def test_removing_statefulset_with_name(cl, client, has_conf):
     remove_statefulset("mystatefulset")
 
     v1.list_namespaced_stateful_set.assert_called_with(
-        "default", field_selector="metadata.name=mystatefulset")
+        "default", field_selector="metadata.name=mystatefulset"
+    )
     assert v1.delete_namespaced_stateful_set.call_count == 1
     v1.delete_namespaced_stateful_set.assert_called_with(
-        "mystatefulset", "default", body=ANY)
+        "mystatefulset", "default", body=ANY
+    )
 
 
-@patch('chaosk8s.has_local_config_file', autospec=True)
-@patch('chaosk8s.statefulset.actions.client', autospec=True)
-@patch('chaosk8s.client')
+@patch("chaosk8s.has_local_config_file", autospec=True)
+@patch("chaosk8s.statefulset.actions.client", autospec=True)
+@patch("chaosk8s.client")
 def test_removing_statefulset_with_label_selector(cl, client, has_conf):
     has_conf.return_value = False
 
@@ -68,15 +73,17 @@ def test_removing_statefulset_with_label_selector(cl, client, has_conf):
     remove_statefulset(label_selector=label_selector)
 
     v1.list_namespaced_stateful_set.assert_called_with(
-        "default", label_selector=label_selector)
+        "default", label_selector=label_selector
+    )
     assert v1.delete_namespaced_stateful_set.call_count == 1
     v1.delete_namespaced_stateful_set.assert_called_with(
-        "mystatefulset", "default", body=ANY)
+        "mystatefulset", "default", body=ANY
+    )
 
 
-@patch('chaosk8s.has_local_config_file', autospec=True)
-@patch('chaosk8s.statefulset.actions.client', autospec=True)
-@patch('chaosk8s.client')
+@patch("chaosk8s.has_local_config_file", autospec=True)
+@patch("chaosk8s.statefulset.actions.client", autospec=True)
+@patch("chaosk8s.client")
 def test_creating_statefulset_with_file_json(cl, client, has_conf):
     has_conf.return_value = False
 
@@ -88,13 +95,12 @@ def test_creating_statefulset_with_file_json(cl, client, has_conf):
     create_statefulset("tests/fixtures/statefulset/create/file.json")
 
     assert v1.create_namespaced_stateful_set.call_count == 1
-    v1.create_namespaced_stateful_set.assert_called_with(
-        "default", body=body)
+    v1.create_namespaced_stateful_set.assert_called_with("default", body=body)
 
 
-@patch('chaosk8s.has_local_config_file', autospec=True)
-@patch('chaosk8s.statefulset.actions.client', autospec=True)
-@patch('chaosk8s.client')
+@patch("chaosk8s.has_local_config_file", autospec=True)
+@patch("chaosk8s.statefulset.actions.client", autospec=True)
+@patch("chaosk8s.client")
 def test_creating_statefulset_with_file_yaml(cl, client, has_conf):
     has_conf.return_value = False
 
@@ -106,13 +112,12 @@ def test_creating_statefulset_with_file_yaml(cl, client, has_conf):
     create_statefulset("tests/fixtures/statefulset/create/file.yaml")
 
     assert v1.create_namespaced_stateful_set.call_count == 1
-    v1.create_namespaced_stateful_set.assert_called_with(
-        "default", body=body)
+    v1.create_namespaced_stateful_set.assert_called_with("default", body=body)
 
 
-@patch('chaosk8s.has_local_config_file', autospec=True)
-@patch('chaosk8s.statefulset.actions.client', autospec=True)
-@patch('chaosk8s.client')
+@patch("chaosk8s.has_local_config_file", autospec=True)
+@patch("chaosk8s.statefulset.actions.client", autospec=True)
+@patch("chaosk8s.client")
 def test_creating_statefulset_with_file_txt_KO(cl, client, has_conf):
     has_conf.return_value = False
 
@@ -123,4 +128,4 @@ def test_creating_statefulset_with_file_txt_KO(cl, client, has_conf):
 
     with pytest.raises(ActivityFailed) as excinfo:
         create_statefulset(path)
-    assert "cannot process {path}".format(path=path) in str(excinfo.value)
+    assert f"cannot process {path}" in str(excinfo.value)
